+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 * Update icon font
 
+**Refactor:**
+
+* Fix method calls from templates
+
 ## 16.3.0 (May 08, 2019)
 
 **New features:**